+++ src/lib.rs
@@ -30,6 +30,12 @@ extern crate strum;
 #[macro_use]
 extern crate strum_macros;
 
+#[cfg(feature="serialize")]
+extern crate serde;
+
+#[cfg(feature="serialize")]
+extern crate serde_json;
+
 use std::collections::HashMap;
 use std::convert;
 use std::net;